Dismantle kernel mod, make Kernel fields private
