+++ radix-engine/src/blueprints/access_controller/executables.rs
@@ -1,9 +1,11 @@
 use super::state_machine::*;
 use super::*;
 use crate::errors::{ApplicationError, RuntimeError};
-use crate::kernel::{
-    deref_and_update, CallFrameUpdate, ExecutableInvocation, Executor, KernelNodeApi,
-    KernelSubstateApi, LockFlags, ResolvedActor,
+use crate::kernel::actor::ResolvedActor;
+use crate::kernel::call_frame::CallFrameUpdate;
+use crate::kernel::interpreters::deref_and_update;
+use crate::kernel::kernel_api::{
+    ExecutableInvocation, Executor, KernelNodeApi, KernelSubstateApi, LockFlags,
 };
 use crate::system::global::GlobalAddressSubstate;
 use crate::system::node::{RENodeInit, RENodeModuleInit};
