+++ radix-engine/src/blueprints/identity/identity.rs
@@ -1,6 +1,10 @@
 use crate::errors::RuntimeError;
+use crate::kernel::actor::ResolvedActor;
+use crate::kernel::call_frame::CallFrameUpdate;
+use crate::kernel::kernel_api::ExecutableInvocation;
+use crate::kernel::kernel_api::Executor;
+use crate::kernel::kernel_api::KernelNodeApi;
 use crate::kernel::kernel_api::KernelSubstateApi;
-use crate::kernel::*;
 use crate::system::global::GlobalAddressSubstate;
 use crate::system::node::RENodeInit;
 use crate::system::node::RENodeModuleInit;
