+++ radix-engine/src/errors.rs
@@ -6,8 +6,9 @@ use crate::blueprints::resource::{
 };
 use crate::blueprints::transaction_processor::TransactionProcessorError;
 use crate::blueprints::transaction_runtime::TransactionRuntimeError;
+use crate::kernel::actor::{ExecutionMode, ResolvedActor};
 use crate::kernel::kernel_api::LockFlags;
-use crate::kernel::{ExecutionMode, ResolvedActor, TrackError};
+use crate::kernel::track::TrackError;
 use crate::system::kernel_modules::auth::AuthError;
 use crate::system::kernel_modules::costing::CostingError;
 use crate::system::kernel_modules::node_move::NodeMoveError;
