+++ radix-engine/src/system/invocation/resolve_method.rs
@@ -1,8 +1,7 @@
+use crate::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi, LockFlags};
 use crate::{
     blueprints::transaction_processor::TransactionProcessorError,
     errors::{ApplicationError, RuntimeError},
-    kernel::kernel_api::LockFlags,
-    kernel::{kernel_api::KernelSubstateApi, KernelNodeApi},
     types::*,
 };
 use radix_engine_interface::api::{
