+++ radix-engine/src/system/kernel_modules/auth/auth_module.rs
@@ -4,8 +4,13 @@ use super::method_authorization::MethodAuthorizationError;
 use crate::blueprints::clock::Clock;
 use crate::blueprints::epoch_manager::EpochManager;
 use crate::errors::*;
+use crate::kernel::actor::ResolvedActor;
+use crate::kernel::actor::ResolvedReceiver;
+use crate::kernel::call_frame::CallFrameUpdate;
+use crate::kernel::call_frame::RENodeVisibilityOrigin;
+use crate::kernel::kernel_api::KernelModuleApi;
 use crate::kernel::kernel_api::LockFlags;
-use crate::kernel::*;
+use crate::kernel::module::KernelModule;
 use crate::system::node::RENodeInit;
 use crate::system::node_modules::auth::AuthZoneStackSubstate;
 use crate::types::*;
